Remove orchestrator http server
Orchestrator had a POST route that was used to submit a solution, but
now ORM has a /submit route that places the submissionID on a queue.
Since orchestrator was already reading from this queue to handle
submissions and the actual listening of the server route was never
executed, we can remove everything related with this route.

This also changes the way orchestrator calls verdict, sending only a
submission ID instead of the whole submission object.
